Fixed returning Entity from controller. Fixed naming of APIs. Refactored service layer, removed a lot of lambda usage.

+++ src/main/java/com/example/bank/models/dtos/AccountDto.java
@@ -3,13 +3,15 @@ package com.example.bank.models.dtos;
 import io.swagger.v3.oas.annotations.media.Schema;
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 public class AccountDto {
     @Schema(description = "Bank account number", example = "1")
-    private long accountNum;
+    private Long accountNum;
 
     @Schema(description = "Current balance of bank account")
-    private long currentBalance;
+    private BigDecimal currentBalance;
 
     @Schema(description = "Holder of bank account")
     private ClientDto client;
