+++ src/main/java/com/example/bank/models/dtos/PutMoneyToAccountDto.java
@@ -6,6 +6,8 @@ import jakarta.validation.constraints.Min;
 import jakarta.validation.constraints.NotNull;
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 import static io.swagger.v3.oas.annotations.media.Schema.RequiredMode.REQUIRED;
 
 @Data
@@ -13,11 +15,11 @@ public class PutMoneyToAccountDto {
 
     @Schema(description = "Bank account number", example = "1", requiredMode = REQUIRED)
     @NotNull(message = "Account number is mandatory")
-    private long accountNum;
+    private Long accountNum;
 
     @Schema(description = "Amount of money to put on the bank account", minimum = "0", example = "1", requiredMode = REQUIRED)
     @NotNull
     @Min(0)
     @Max(Long.MAX_VALUE)
-    private long amount;
+    private BigDecimal amount;
 }
